-Reconfigured Output compare properly (works without TMR2) -All the CAN operations are now done inside can_app -Added pos,vel& torque sending at ~50hz

+++ Mplab/AGV_Driver_V1.X/motor_driver.c
@@ -12,18 +12,19 @@ void setDuty(float duty)
 	else if(duty>0)
 	{
 		OC2R = 0; //Turn off reverse half-bridge
-		OC1R = limitPWM((float)MAX_PWM_DUTY*(duty/100));
-			}
-			else if(duty<0)
-			{
+		OC1R = (uint16_t) limitPWM((float)MAX_PWM_DUTY*((float)duty/100));
+
+	}
+	else if(duty<0)
+	{
 		OC1R = 0; //Turn off forward half-bridge
-		OC2R = limitPWM((float)MAX_PWM_DUTY*(duty/100));
+		OC2R = (uint16_t) limitPWM((float)MAX_PWM_DUTY*((float)duty/100));
 	}
 }
 
 
 
-float limitPWM(float val){
+uint16_t limitPWM(float val){
 	if (val>MAX_PWM_DUTY_ALLOWED)
 	{
 		return MAX_PWM_DUTY_ALLOWED;
